backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
python-dotenv
email-validator
